requirements.txt
requests
transformers>=4.30.0
torch
sentencepiece
python-dotenv